moved removalrequestrunner to utils/removal_utils.py
